Changed multi platform targets to one universal target
Deleted the tvOS target. Renamed the ios target to Universal.
Added tvOS and macOS to the Universal taget and added automatic
install to the current users Frameworks folder for each platform,
i.e. ~/Library/Frameworks/iphoneos
Changed the bundle name back to the original spelling so it will
work with the flatbuffer compiler